# Conflicts:
#	TestProjectCesiProgSys/obj/Debug/net7.0/TestProjectCesiProgSys.assets.cache
#	TestProjectCesiProgSys/obj/Debug/net7.0/TestProjectCesiProgSys.csproj.AssemblyReference.cache
#	TestProjectCesiProgSys/obj/TestProjectCesiProgSys.csproj.nuget.g.props
#	TestProjectCesiProgSys/obj/project.nuget.cache

+++ CesiProgSys/ToolsBox/Info.cs
@@ -152,8 +152,8 @@ public sealed class Info : INotifyPropertyChanged
                 }
             }
         }
-        private float _progression;
-        public float Progression
+        private int _progression;
+        public int Progression
         {
             get => _progression;
             set
